Add cubozoa bucket and make fish buckets retain variant
This change adds a cubozoa bucket that functions like the other types of
fish buckets. In addition to that, this commit also saves the size and
variant of end fish and cubozoa in the item tag when they are scooped up
in a bucket, so that they retain their size/variant when they are placed
again. Here is the issue from the Forge port that made me away of the
issue: https://github.com/Beethoven92/BetterEndForge/issues/93